Update Pvwatts to work with version 5 of the api
This commit is a modernization of the gem compatible with Version 5 of
NREL's PVWatts api.

Implementation Notes:

Only two additional runtime dependencies were deemed necessary:

 - Virtus used to manage domain objects required for the api request and
   returned in the API response. These are intended to help enforce the
required data types needed by the API, along with assist developers in
understanding which attributes are required.
- Addressable is used to craft RFC 3986, RFC 3987, and RFC 6570 (level
  4) compliant query parameters from the potentially variable inputs
needed to interact with the NREL API.

Development dependencies:

- RSpec/VCR chosen to handle the storage of NREL api requests. However,
  I need to setup CI and include the mechanism for handling the API key
locally and in CI securely and easily for other developers.
- Pry was also included because it's awesome.

The README was updated to provide what I hope is a straightforward
introduction to using the gem. I plan to include a sample app in which
to showcase additional features I have in mind at a later time.